feat: moved FilteredInput into PageTable for SSR and performance reasons
