+++ app/features/admin/pagetable/index.ts
@@ -1,2 +1,1 @@
-export * from "./FilterInput";
 export * from "./PageTable";
